# src/ClientApp/src/typescript-fetch/models/Equipment.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Open API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    InspectionItem,
    InspectionItemFromJSON,
    InspectionItemFromJSONTyped,
    InspectionItemToJSON,
} from './InspectionItem';

/**
 * 
 * @export
 * @interface Equipment
 */
export interface Equipment {
    /**
     * 
     * @type {number}
     * @memberof Equipment
     */
    equipmentId: number;
    /**
     * 
     * @type {number}
     * @memberof Equipment
     */
    orderIndex: number;
    /**
     * 
     * @type {string}
     * @memberof Equipment
     */
    equipmentName: string;
    /**
     * 
     * @type {Array<InspectionItem>}
     * @memberof Equipment
     */
    inspectionItems: Array<InspectionItem>;
}

export function EquipmentFromJSON(json: any): Equipment {
    return EquipmentFromJSONTyped(json, false);
}

export function EquipmentFromJSONTyped(json: any, ignoreDiscriminator: boolean): Equipment {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'equipmentId': json['equipment_id'],
        'orderIndex': json['order_index'],
        'equipmentName': json['equipment_name'],
        'inspectionItems': ((json['inspection_items'] as Array<any>).map(InspectionItemFromJSON)),
    };
}

export function EquipmentToJSON(value?: Equipment | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'equipment_id': value.equipmentId,
        'order_index': value.orderIndex,
        'equipment_name': value.equipmentName,
        'inspection_items': ((value.inspectionItems as Array<any>).map(InspectionItemToJSON)),
    };
}