# src/ClientApp/src/typescript-fetch/models/InspectionSheet.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Open API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    Equipment,
    EquipmentFromJSON,
    EquipmentFromJSONTyped,
    EquipmentToJSON,
} from './Equipment';

/**
 * 
 * @export
 * @interface InspectionSheet
 */
export interface InspectionSheet {
    /**
     * 
     * @type {number}
     * @memberof InspectionSheet
     */
    sheetId: number;
    /**
     * 
     * @type {string}
     * @memberof InspectionSheet
     */
    sheetName: string;
    /**
     * 
     * @type {number}
     * @memberof InspectionSheet
     */
    inspectionTypeId: number;
    /**
     * 
     * @type {number}
     * @memberof InspectionSheet
     */
    inspectionGroupId: number;
    /**
     * 
     * @type {Array<Equipment>}
     * @memberof InspectionSheet
     */
    equipments: Array<Equipment>;
}

export function InspectionSheetFromJSON(json: any): InspectionSheet {
    return InspectionSheetFromJSONTyped(json, false);
}

export function InspectionSheetFromJSONTyped(json: any, ignoreDiscriminator: boolean): InspectionSheet {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'sheetId': json['sheet_id'],
        'sheetName': json['sheet_name'],
        'inspectionTypeId': json['inspection_type_id'],
        'inspectionGroupId': json['inspection_group_id'],
        'equipments': ((json['equipments'] as Array<any>).map(EquipmentFromJSON)),
    };
}

export function InspectionSheetToJSON(value?: InspectionSheet | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'sheet_id': value.sheetId,
        'sheet_name': value.sheetName,
        'inspection_type_id': value.inspectionTypeId,
        'inspection_group_id': value.inspectionGroupId,
        'equipments': ((value.equipments as Array<any>).map(EquipmentToJSON)),
    };
}